requirements.txt
PyExifTool>=0.5.0
pytest>=7.0.0
Pillow>=9.0.0
